Fix Windows /W3 warnings
git-svn-id: svn+ssh://svn.code.sf.net/p/x3270/x3270/trunk@1157 12e4f418-770b-42d6-82dd-bd6eefd7344d

+++ x3270/Common/XtGlue.c
@@ -682,7 +682,7 @@ process_events(Boolean block)
 			if (now > timeouts->ts)
 				tmo = 0;
 			else
-				tmo = timeouts->ts - now;
+				tmo = (DWORD)(timeouts->ts - now);
 #else /*][*/
 			(void) gettimeofday(&now, (void *)NULL);
 			twait.tv_sec = timeouts->tv.tv_sec - now.tv_sec;
